requirements.txt
numpy
opencv-python>=4.5.4
websockets
Pillow